requirements.txt
hexaly==14.0.20250814
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
requests==2.31.0
